Finish splitting out semantics_handle.cpp
