+++ toolchain/semantics/semantics_handle_function.cpp
@@ -94,4 +94,14 @@ auto SemanticsHandleFunctionIntroducer(SemanticsContext& context,
   return true;
 }
 
+auto SemanticsHandleReturnType(SemanticsContext& context,
+                               ParseTree::Node parse_node) -> bool {
+  // Propagate the type expression.
+  auto [type_parse_node, type_node_id] =
+      context.node_stack().PopWithParseNode<SemanticsNodeId>();
+  auto cast_node_id = context.ExpressionAsType(type_parse_node, type_node_id);
+  context.node_stack().Push(parse_node, cast_node_id);
+  return true;
+}
+
 }  // namespace Carbon
