+++ toolchain/semantics/semantics_handle_noop.cpp
@@ -1,0 +1,26 @@
+// Part of the Carbon Language project, under the Apache License v2.0 with LLVM
+// Exceptions. See /LICENSE for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#include "toolchain/semantics/semantics_context.h"
+
+namespace Carbon {
+
+auto SemanticsHandleEmptyDeclaration(SemanticsContext& /*context*/,
+                                     ParseTree::Node /*parse_node*/) -> bool {
+  // Empty declarations have no actions associated.
+  return true;
+}
+
+auto SemanticsHandleFileEnd(SemanticsContext& /*context*/,
+                            ParseTree::Node /*parse_node*/) -> bool {
+  // Do nothing, no need to balance this node.
+  return true;
+}
+
+auto SemanticsHandleInvalidParse(SemanticsContext& context,
+                                 ParseTree::Node parse_node) -> bool {
+  return context.TODO(parse_node, "HandleInvalidParse");
+}
+
+}  // namespace Carbon
